fix(#44): remove false signature verification and improve date filtering
- Removed X-Strava-Signature check from ingest Lambda
- Improved date filtering in forecast-api to handle ISO timestamps
- Refactored API client for better testability
- Added unit and integration tests
- Updated CDK stack to use Regional API Gateway
- Updated deploy script to load STRAVA.secrets automatically
