+++ apps/strava-webhook/package.json
@@ -13,7 +13,7 @@
     "test": "vitest",
     "cdk": "cdk",
     "synth": "cdk synth",
-    "deploy": "cdk deploy",
+    "deploy": "if [ ! -f ../../STRAVA.secrets ]; then echo 'Error: STRAVA.secrets not found in root'; exit 1; fi; . ../../STRAVA.secrets && cdk deploy",
     "diff": "cdk diff",
     "destroy": "cdk destroy"
   },
